requirements.txt
# Sift — requirements
# Python 3.11+

# UI / tray
rumps>=0.4.0; sys_platform == "darwin"
pystray>=0.19.5; sys_platform != "darwin"
Pillow>=10.0.0

# ML
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=3.0.0
accelerate>=0.30.0

# Clustering / ranking
scikit-learn>=1.4.0
numpy>=1.26.0

# ArXiv — using requests + feedparser directly for timeout/header control
requests>=2.32.0
feedparser>=6.0.0

# Scheduling
APScheduler>=3.10.4

# Config / templating
PyYAML>=6.0.1
Jinja2>=3.1.3

# Paths
platformdirs>=4.2.0

# System info
psutil>=5.9.8

# macOS native notifications (optional, graceful fallback)
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
pyobjc-framework-WebKit>=10.0; sys_platform == "darwin"
